respect 32 bit ops

+++ src/n64/instr_r.cpp
@@ -53,19 +53,19 @@ void instr_slt(N64 &n64, const Opcode &opcode)
 void instr_subu(N64 &n64, const Opcode &opcode)
 {
     // does not trap on overflow
-    n64.cpu.regs[opcode.rd] = sign_extend_mips<s64,s32>(n64.cpu.regs[opcode.rs] - n64.cpu.regs[opcode.rt]);
+    n64.cpu.regs[opcode.rd] = sign_extend_mips<s64,s32>(u32(n64.cpu.regs[opcode.rs]) - u32(n64.cpu.regs[opcode.rt]));
 }
 
 void instr_addu(N64 &n64, const Opcode &opcode)
 {
     // does not trap on overflow
-    n64.cpu.regs[opcode.rd] = sign_extend_mips<s64,s32>(n64.cpu.regs[opcode.rs] + n64.cpu.regs[opcode.rt]);
+    n64.cpu.regs[opcode.rd] = sign_extend_mips<s64,s32>(u32(n64.cpu.regs[opcode.rs]) + u32(n64.cpu.regs[opcode.rt]));
 }
 
 void instr_add(N64 &n64, const Opcode &opcode)
 {
     // does not trap on overflow
-    const auto ans =  sign_extend_mips<s64,s32>(n64.cpu.regs[opcode.rs] + n64.cpu.regs[opcode.rt]);
+    const auto ans = sign_extend_mips<s64,s32>(u32(n64.cpu.regs[opcode.rs]) + u32(n64.cpu.regs[opcode.rt]));
 
     if(did_overflow(n64.cpu.regs[opcode.rs],n64.cpu.regs[opcode.rt],ans))
     {
@@ -88,7 +88,7 @@ void instr_and(N64 &n64, const Opcode &opcode)
 
 void instr_multu(N64 &n64, const Opcode &opcode)
 {
-    const u64 res = static_cast<u32>(n64.cpu.regs[opcode.rs]) * static_cast<u32>(n64.cpu.regs[opcode.rt]);
+    const u64 res = u32(n64.cpu.regs[opcode.rs]) * u32(n64.cpu.regs[opcode.rt]);
 
     n64.cpu.lo = sign_extend_mips<s64,s32>(res & 0xffffffff);
     n64.cpu.hi = sign_extend_mips<s64,s32>((res >> 32) & 0xffffffff);
